refactor: GameService 패키지 이동
