+++ src/test/java/chess/dao/GameServiceTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.dto.BoardDto;
 import chess.model.position.Position;
+import chess.service.GameService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
